added tests

+++ src/main/kotlin/Card.kt
@@ -1,6 +1,6 @@
 package ru.netology
 
-class Card constructor(var cardType: String, var transactedLastMonth: Int, var sumToTransact: Int, var sumCard: Int) {
+class Card constructor(var cardType: String, var transactedLastMonth: Int, var sumToTransact: Double, var sumCard: Double) {
 
 
 }
